# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: webcup
  description: Best API REST building with GoaDesign
  contact:
    name: Guillaume MORIN
    email: guillaume.morin@epitech.eu
    url: https://guillaume-morin.fr/
  license:
    name: GOA X GM
    url: https://guillaume-morin.fr/
  version: "1.0"
host: :0
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /v1/web/user/{id}:
    get:
      tags:
      - users
      summary: getUserByID users
      description: Get one User
      operationId: users#getUserByID
      parameters:
      - name: id
        in: path
        description: Unique ID of the User
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersGetUserByIDResponseBody'
            required:
            - user
            - success
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UsersGetUserByIDUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
  /v1/web/user/edit:
    put:
      tags:
      - users
      summary: updateDescription users
      description: Update main info like email, username, firstname, lastname
      operationId: users#updateDescription
      parameters:
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: UpdateDescriptionRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersUpdateDescriptionRequestBody'
          required:
          - email
          - id
          - username
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersUpdateDescriptionResponseBody'
            required:
            - success
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UsersUpdateDescriptionUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
  /v1/web/user/edit/avatar:
    put:
      tags:
      - users
      summary: updateAvatar users
      description: Update avatar
      operationId: users#updateAvatar
      parameters:
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: UpdateAvatarRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersUpdateAvatarRequestBody'
          required:
          - id
          - avatar
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersUpdateAvatarResponseBody'
            required:
            - success
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UsersUpdateAvatarUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
  /v1/web/user/edit/stock:
    put:
      tags:
      - users
      summary: updateNumberStockage users
      description: Update number stockage
      operationId: users#updateNumberStockage
      parameters:
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: UpdateNumberStockageRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersUpdateNumberStockageRequestBody'
          required:
          - id
          - number
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersUpdateNumberStockageResponseBody'
            required:
            - success
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UsersUpdateNumberStockageUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
  /v1/web/user/remove/{id}:
    delete:
      tags:
      - users
      summary: deleteUser users
      description: Delete one User by ID
      operationId: users#deleteUser
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        description: OAuth token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      - name: jwtToken
        in: header
        description: Jwt token
        required: false
        type: string
        pattern: ^Bearer [^ ]+$
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersDeleteUserResponseBody'
            required:
            - success
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UsersDeleteUserUnknownErrorResponseBody'
            required:
            - err
            - success
            - error_code
      schemes:
      - http
      security:
      - OAuth2_header_Authorization: []
        jwt_header_jwtToken: []
...
definitions:
...
  UsersDeleteUserResponseBody:
    title: UsersDeleteUserResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: false
    example:
      success: false
    required:
    - success
  UsersDeleteUserUnknownErrorResponseBody:
    title: UsersDeleteUserUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: true
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: false
    required:
    - err
    - success
    - error_code
  UsersGetUserByIDResponseBody:
    title: UsersGetUserByIDResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: false
      user:
        $ref: '#/definitions/resUserResponseBody'
    example:
      success: false
      user:
        avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
    required:
    - user
    - success
  UsersGetUserByIDUnknownErrorResponseBody:
    title: UsersGetUserByIDUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: false
    required:
    - err
    - success
    - error_code
  UsersUpdateAvatarRequestBody:
    title: UsersUpdateAvatarRequestBody
    type: object
    properties:
      avatar:
        type: string
        description: Url of the avatar and stock in db
        example: Veritatis voluptas beatae quibusdam fugit.
      id:
        type: string
        example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        format: uuid
    example:
      avatar: Itaque a veniam ut perspiciatis officiis.
      id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
    required:
    - id
    - avatar
  UsersUpdateAvatarResponseBody:
    title: UsersUpdateAvatarResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: false
      user:
        $ref: '#/definitions/resUserResponseBody'
    example:
      success: true
      user:
        avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
    required:
    - success
  UsersUpdateAvatarUnknownErrorResponseBody:
    title: UsersUpdateAvatarUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: true
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: true
    required:
    - err
    - success
    - error_code
  UsersUpdateDescriptionRequestBody:
    title: UsersUpdateDescriptionRequestBody
    type: object
    properties:
      email:
        type: string
        example: guillaume@gmail.com
        format: email
      firstname:
        type: string
        example: Guillaume
        minLength: 3
        maxLength: 20
      id:
        type: string
        example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        format: uuid
      lastname:
        type: string
        example: Morin
        minLength: 3
        maxLength: 20
      username:
        type: string
        example: guillaumemoriin
        pattern: ^[a-z0-9_\-]+$
        minLength: 3
        maxLength: 20
    example:
      email: guillaume@gmail.com
      firstname: Guillaume
      id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      lastname: Morin
      username: guillaumemoriin
    required:
    - email
    - id
    - username
  UsersUpdateDescriptionResponseBody:
    title: UsersUpdateDescriptionResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: false
      user:
        $ref: '#/definitions/resUserResponseBody'
    example:
      success: true
      user:
        avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
        email: guillaume@gmail.com
        firstname: Guillaume
        id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        lastname: Morin
        role: Maiores et non dolor facilis et voluptatem.
        username: guillaumemorin
    required:
    - success
  UsersUpdateDescriptionUnknownErrorResponseBody:
    title: UsersUpdateDescriptionUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: false
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: false
    required:
    - err
    - success
    - error_code
  UsersUpdateNumberStockageRequestBody:
    title: UsersUpdateNumberStockageRequestBody
    type: object
    properties:
      id:
        type: string
        example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
        format: uuid
      number:
        type: integer
        example: 5128633245619362707
        format: int64
    example:
      id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      number: 2194423627104278954
    required:
    - id
    - number
  UsersUpdateNumberStockageResponseBody:
    title: UsersUpdateNumberStockageResponseBody
    type: object
    properties:
      success:
        type: boolean
        example: true
    example:
      success: true
    required:
    - success
  UsersUpdateNumberStockageUnknownErrorResponseBody:
    title: UsersUpdateNumberStockageUnknownErrorResponseBody
    type: object
    properties:
      err:
        type: string
        example: sql no rows affected
      error_code:
        type: string
        example: TX_UPDATE_ITEM
      success:
        type: boolean
        default: false
        example: true
    description: Error not identified (500)
    example:
      err: sql no rows affected
      error_code: TX_UPDATE_ITEM
      success: true
    required:
    - err
    - success
    - error_code
...
  resUserResponseBody:
    title: resUserResponseBody
    type: object
    properties:
      avatar:
        type: string
        example: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
      email:
        type: string
        example: guillaume@gmail.com
      firstname:
        type: string
        example: Guillaume
      id:
        type: string
        example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      lastname:
        type: string
        example: Morin
      role:
        type: string
        description: User is admin or not
        example: Vitae et.
      username:
        type: string
        example: guillaumemorin
    example:
      avatar: /public/uploads/2022/02/eedf427a-559a-4faf-9e75-357fbb5d65fb.png
      email: guillaume@gmail.com
      firstname: Guillaume
      id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
      lastname: Morin
      role: Et distinctio nihil.
      username: guillaumemorin
    required:
    - id
    - email
    - username
    - firstname
    - lastname
    - avatar
    - role
securityDefinitions:
  OAuth2_header_Authorization:
    type: oauth2
    description: Use OAuth2 to authenticate
    flow: application
    tokenUrl: /authorization
    scopes:
      api:read: Provides read access
  jwt_header_jwtToken:
    type: apiKey
    description: |-
      Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

      **Security Scopes**:
        * `api:read`: Read-only access
        * `api:write`: Read and write access
    name: jwtToken
    in: header